+++ Endpoints/Produtos/ProdutoDelete.cs
@@ -5,13 +5,17 @@ namespace LojaAPI.Endpoints.Produtos;
 
 public class ProdutoDelete
 {
-    public static string Template => "/produtos/{id}";
+    public static string Template => "/produtos/{id:guid}";
     public static string[] Methods => new string[] { HttpMethod.Delete.ToString() };
     public static Delegate Handle => Action;
 
     public static IResult Action([FromRoute] Guid Id, ApplicationDbContext context)
     {
-        var produto = context.Produtos.Where(p => p.Id == Id).First();
+        var produto = context.Produtos.Where(p => p.Id == Id).FirstOrDefault();
+        if (produto == null)
+        {
+            return Results.NotFound("Produto não existe no Banco de Dados");
+        }
         context.Produtos.Remove(produto);
         context.SaveChanges();
         return Results.Ok();
